apps/uptime-kuma/requirements.txt
uptime-kuma-api-v2==1.0.1
PyYAML==6.0.3
